# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DigiSchool API",
        "description": "API REST pour la gestion d'ecoles DigiSchool.\n\n## Authentification\nLa plupart des endpoints necessitent un token JWT.\n1. Utilisez `POST /api/auth/login` pour obtenir un token\n2. Cliquez sur le bouton \"Authorize\" et entrez: `Bearer <votre_token>`\n\n## Multi-tenancy\nLe systeme est multi-tenant base sur l'ecole (ecoleId).\nL'ecoleId est automatiquement extrait du token JWT.\n\n## Comptes de test\n| Role | Email | Telephone | Password |\n|------|-------|-----------|----------|\n| Admin | admin@digischool.cm | +237600000000 | Admin@2025 |\n| Directeur | smbarga@lavictoire.cm | +237677123456 | Directeur@2025 |\n| Enseignant | jpkamga@lavictoire.cm | +237670111222 | Enseignant@2025 |\n| Secretaire | catangana@lavictoire.cm | +237660555666 | Secretaire@2025 |\n| Parent | fnkoulou@gmail.com | +237691666777 | Parent@2025 |\n",
        "contact": {
            "name": "DigiSchool Support",
            "url": "https://digischool.cm",
            "email": "support@digischool.cm"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://digischool.cm/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Serveur de developpement local"
        },
        {
            "url": "https://api.digischool.cm",
...
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/api/regions/{id}": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getById_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "update_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/quartiers/{id}": {
            "get": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "getById_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuartierDto"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "update_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuartierDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuartierDto"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "delete_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/regions": {
            "get": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "getAll_2",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RegionDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "region-controller"
                ],
                "operationId": "create_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/quartiers": {
            "get": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "getAll_3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuartierDto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "create_3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/QuartierDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/QuartierDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/quartiers/ville/{villeId}": {
            "get": {
                "tags": [
                    "quartier-controller"
                ],
                "operationId": "getByVilleId",
                "parameters": [
                    {
                        "name": "villeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuartierDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/departements/region/{regionId}": {
            "get": {
                "tags": [
                    "departement-controller"
                ],
                "operationId": "getByRegionId",
                "parameters": [
                    {
                        "name": "regionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DepartementDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/adresses/quartier/{quartierId}": {
            "get": {
                "tags": [
                    "adresse-controller"
                ],
                "operationId": "getByQuartierId",
                "parameters": [
                    {
                        "name": "quartierId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AdresseDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "RegionDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    }
                }
            },
            "QuartierDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "villeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "villeNom": {
                        "type": "string"
                    }
                }
            },
...
            "DepartementDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "code": {
                        "type": "string"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "chefLieu": {
                        "type": "string"
                    },
                    "regionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "regionNom": {
                        "type": "string"
                    }
                }
            },
...
            "AdresseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quartierId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "quartierNom": {
                        "type": "string"
                    },
                    "rue": {
                        "type": "string"
                    },
                    "repere": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via POST /api/auth/login. Entrez: Bearer <token>",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}